UI/requirements.txt
tensorflow-gpu
Keras
numpy
opencv-python
scikit-image
SimpleITK
PyInstaller
